Move the `Type1Font` from `src/core/fonts.js` and into its own file
